rename FilterEvents to FilteredEvents
