+++ src/app/events/page.tsx
@@ -1,22 +1,20 @@
 "use client";
 
 import type { NextPage } from "next";
+import { useSession } from "next-auth/react";
+
+import { EventFilters } from "@/components/widgets/EventFilters";
+import { CreateAndInvite } from "@/components/widgets/CreateAndInvite";
+import { Mode } from "@/components/widgets/CreateAndInvite/CreateAndInvite";
 
 import { Title } from "@/components/shared/Title";
 import { Slider } from "@/components/shared/Slider";
-import { EventFilters } from "@/components/widgets/EventFilters";
-
-import { CreateAndInvite } from "@/components/widgets/CreateAndInvite";
-import { useSession } from "next-auth/react";
-import { Mode } from "@/components/widgets/CreateAndInvite/CreateAndInvite";
 
 import classes from "./page.module.css";
 
 const EventListPage: NextPage = () => {
     const { status } = useSession();
-    //TODO: реализовать поиск внутри 4-х угольника
 
-    //TODO: добавить индекс поиска в mongoDB - db.events.createIndex({ location: "2dsphere" });
     return (
         <div className={classes.eventListPage}>
             <Title tag={"h1"} title="Join the adventure" className={classes.title} />
