--- src/components/widgets/EventFilters/FilterEvents/index.ts
@@ -1,1 +1,0 @@
-export { FilterEvents } from "./FilterEvents";
