+++ src/components/widgets/EventFilters/FilteredEvents/FilteredEvents.module.css
