+++ src/components/widgets/EventFilters/FilteredEvents/index.ts
@@ -1,0 +1,1 @@
+export { FilteredEvents } from "./FilteredEvents";
